requirements.txt
Click==7.0
Flask==1.0.2
itsdangerous==1.1.0
Werkzeug==0.14.1
SQLAlchemy==1.2.17
Flask-SQLAlchemy==2.3.2
Flask-RESTful==0.3.7
Flask-Cors==3.0.7
Jinja2==2.10
MarkupSafe==1.1.0
six==1.12.0
python-dateutil==2.7.5
python-dotenv==0.10.1
python-memcached==1.59
sendgrid==5.6.0
firebase-admin==2.16.0
google-cloud-firestore==0.32.1
google-cloud-storage==1.15.0
#psycopg2-binary==2.7.7
xlrd==1.0.0
